extract startgame page
