+++ src/common.rs
@@ -17,3 +17,9 @@ pub mod tempo_const {
     pub const AUTO_ATK_TEMPO_MS: u64 = 3000;
     pub const TIMER_FUTURE_1S: u64 = 1000;
 }
+
+#[derive(Debug, Clone, PartialEq)]
+pub enum ButtonStatus {
+    StartGame = 0,
+    ReplayGame,
+}
