+++ src/components/mod.rs
@@ -1,2 +1,3 @@
 pub mod character_page;
 pub mod gameboard;
+pub mod startgame_page;
